+++ test-core/matrix/main_test_matrix.cpp
@@ -1676,7 +1676,7 @@ TEST(SparseMatrix, denseCols2)
     auto D = F.denseCols(0, 0, 1, 2, rowIdx);
     ASSERT_EQ(D.w, 1);
     ASSERT_EQ(D.h, 1);
-    ASSERT_EQ(rowIdx.size(), 1);
+    ASSERT_EQ(rowIdx.size(), (size_t)1);
     ASSERT_EQ(rowIdx[0], 1);
 }
 
@@ -1692,7 +1692,7 @@ TEST(SparseMatrix, denseRows2)
     auto D = F.denseRows(0, 0, 2, 1, rowIdx);
     ASSERT_EQ(D.w, 1);
     ASSERT_EQ(D.h, 1);
-    ASSERT_EQ(rowIdx.size(), 1);
+    ASSERT_EQ(rowIdx.size(), (size_t)1);
     ASSERT_EQ(rowIdx[0], 1);
 }
 
@@ -1709,7 +1709,7 @@ TEST(SparseMatrix, denseCols3)
     auto D = F.denseCols(1, 0, 2, 2, rowIdx);
     ASSERT_EQ(D.w, 1);
     ASSERT_EQ(D.h, 0);
-    ASSERT_EQ(rowIdx.size(), 0);
+    ASSERT_EQ(rowIdx.size(), (size_t)0);
 }
 
 TEST(SparseMatrix, denseCols4)
@@ -1724,7 +1724,7 @@ TEST(SparseMatrix, denseCols4)
     auto D = F.denseCols(2, 0, 3, 2, rowIdx);
     ASSERT_EQ(D.w, 1);
     ASSERT_EQ(D.h, 1);
-    ASSERT_EQ(rowIdx.size(), 1);
+    ASSERT_EQ(rowIdx.size(), (size_t)1);
     ASSERT_EQ(rowIdx[0], 1);
     ASSERT_EQ(D.a(0, 0), 5.0);
 }
